Add a status check script. Make the test longer so that status will show up.
There is a timing delay when the output.txt file gets sync'ed to the local file system. This seems to be an NFS issue since if you are on the first allocated node then you can see the file update in real time. But since we are accessing the file from the headnode there is a delay when NFS decides to sync it. It's not a big problem at the moment, but something to lookout for.